+++ frontend/src/components/demos/taxtech/TaxDemo.tsx
@@ -1,126 +1,190 @@
-import { Component, ChangeEvent } from 'react';
+import { useState, useEffect } from 'react';
+import {useDispatch, useSelector} from 'react-redux';
 import axios from 'axios';
-import { AxiosError } from 'axios';
 import TextField from '@mui/material/TextField';
-import Button from '@mui/material/Button';
+import {Button, Typography} from '@mui/material';
 import Select from '@mui/material/Select';
 import MenuItem from '@mui/material/MenuItem';
 import LinearProgress from '@mui/material/LinearProgress';
 import Box from '@mui/material/Box';
-
 import { SelectChangeEvent } from '@mui/material/Select';
-import TaxDataDashboard from './TaxDataDashboard';
-import { Grid,Divider} from '@mui/material';
+import { Divider } from '@mui/material';
+import {LocalizationProvider, DatePicker} from '@mui/x-date-pickers';
+import {AdapterDayjs} from '@mui/x-date-pickers/AdapterDayjs';
+import {DataGrid, GridColDef, GridToolbarContainer, useGridApiContext}
+  from '@mui/x-data-grid';
+import dayjs,{ Dayjs } from 'dayjs';
 
-interface AppState {
-  tableAIProgressLoading: boolean;
-  textresponse: any;
-  textprompt: string;
-  isLoading: boolean;
-  models: Map<string, any>; 
-  textSelectedModel: string; 
+import {fetchTaxDashboardData} from '../../../reducers/tax/TaxDashboardDataReducer';
+import {RootState, AppDispatch} from '../../../reducers/store';
+const models = new Map([
+  ['mistral-nemo', 'open-mistral-nemo'],
+  ['mistral-small', 'mistral-small-latest'],
+  ['mistral-large', 'mistral-large-latest'],
+]);
+
+
+interface Message {
+  id: bigint;
+  appointment_day: string;
+  appointment_time: string;
+  doctor_name: string;
+  status: string;
+  observations: string;
 }
 
-type TaxTechDemoProps = {
-  serverUrl: string;
-  isOnline: boolean;
-};
+const columns: GridColDef<Message>[] = [
+  {field: 'id', headerName: 'ID', width: 90},
+  {
+    field: 'appointment_day',
+    headerName: 'Day',
+    width: 150,
+    editable: false,
+  },
+  {
+    field: 'appointment_time',
+    headerName: 'Time',
+    width: 150,
+    editable: false,
+  },
+  {
+    field: 'doctor_name',
+    headerName: 'Doctor',
+    width: 150,
+    editable: false,
+  },
+  {
+    field: 'status',
+    width: 150,
+    headerName: 'Status',
+    editable: false,
+  },
+  {
+    field: 'observations',
+    headerName: 'Observations',
+    width: 150,
+    editable: false,
+  },
+];
+
+/** render
+ * @return {return}
+ */
+function CustomExportButton() {
+  const apiRef = useGridApiContext();
+
+  const handleExport = () => {
+    apiRef.current.exportDataAsCsv();
+  };
+
+  return (
+    <Button onClick={handleExport}>
+      Download CSV
+    </Button>
+  );
+}
+
+/** render
+ * @return {return}
+ */
+function CustomToolbar() {
+  return (
+    <GridToolbarContainer>
+      <CustomExportButton />
+    </GridToolbarContainer>
+  );
+}
+
+const todayDate = new Date().toISOString().slice(0, 10);
+const today = new Date();
+const nextSevenDays = new Date(today.getTime() +
+  (7 * 24 * 60 * 60 * 1000)).toISOString().slice(0, 10);
 
 
-//const [tableAIProgressLoading, setTableAIProgressLoading] = useState<boolean>(false);
-class TaxTechDemo extends Component<TaxTechDemoProps, AppState> {
-  ollamaBaseUrl = import.meta.env.VITE_OLLAMA_BASE_URL;
-  serverBaseUrl = "http://localhost:8000/api/v1" ;
-  isOnline= true;
-  
-  //serverBaseUrl = this.hfBaseUrl; 
-  constructor(props:TaxTechDemoProps) {
-    super(props);
-    //this.serverBaseUrl = this.props.serverUrl;
-    this.isOnline = this.props.isOnline;
-    this.state = {
-      textresponse: null,
-      tableAIProgressLoading: false,
-      textprompt: '',
-      isLoading: false,
-      models: new Map([
-        ['mistral-nemo', 'open-mistral-nemo'],
-        ['mistral-small','mistral-small-latest'],
-        ['mistral-large','mistral-large-latest']
-      ]), 
-      textSelectedModel: 'mistral-nemo',
-    };
-    //console.log(this.hfBaseUrl);
-    //console.log(this.localInferenceUrl);
+const TaxTechDemo: React.FC<{ serverUrl: string; isOnline: boolean }> = ({ serverUrl, isOnline }) => {
 
-  }
+  console.log(serverUrl);
+  const [textPrompt, setTextPrompt] = useState<string>('');
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [textResponse, setTextResponse] = useState<string | null>(null);
+  const [textSelectedModel, setTextSelectedModel] = useState<string>('mistral-nemo');
+  const [tableAIProgressLoading, setTableAIProgressLoading] = useState<boolean>(false);
 
-  componentDidMount() {
-    //this.getOrPullModel(this.state.selectedModel);
-  }
+  const [timerows, setTimerows] = useState<Array<Message>>([]);
+  const [loading, setLoading] = useState(true); // add loading state
 
-  checkModelExists = async (modelName:string) => {
-    try {
-      await axios.post(`${this.ollamaBaseUrl}/show`, { name: modelName });
-      return true; // Model exists
-    } catch (error) {
-      if (error instanceof AxiosError && error.response && error.response.status === 404) {
-        return false; // Model doesn't exist
-      }
-      throw error; // Rethrow other errors
+  const dispatch = useDispatch<AppDispatch>();
+  const [startDate, setStartDate] = useState<Dayjs | null>(null);
+  const [endDate, setEndDate] = useState<Dayjs | null>(null);
+
+  const userId = 1;
+
+  const taxDashboardDataList = useSelector((state: RootState) =>
+    state.taxDashboardDataList.userData);
+
+
+  useEffect(() => {
+    if (loading) {
+    dispatch(
+      fetchTaxDashboardData({
+        page: 1,
+        appointment_day_after: todayDate,
+        appointment_day_before: nextSevenDays,
+        user_id: userId,
+        rejectValue: 'Failed to fetch Appointment.',
+      })
+  ).then(() => setLoading(false));
+}
+}, [dispatch, todayDate, nextSevenDays, userId, loading]);
+
+  useEffect(() => {
+    if (startDate && endDate) {
+      const filteredData = taxDashboardDataList.filter((item:any) => {
+        const itemDate = dayjs(item.appointment_day);
+        return itemDate.isAfter(startDate) && itemDate.isBefore(endDate);
+      });
+      setTimerows(filteredData);
+    } else {
+      setTimerows(taxDashboardDataList);
     }
-  };
+  }, [taxDashboardDataList, startDate, endDate]);
 
 
 
+  const sendPromptToServer = async () => {
+    setTableAIProgressLoading(true);
 
-  handleTextPromptChange = (event: ChangeEvent<HTMLInputElement>) => {
-    this.setState({ textprompt: event.target.value });
-  };
-
-  handleTextModelChange = (event: SelectChangeEvent<string>) => {
-    this.setState({ textSelectedModel: event.target.value }, () => {
-      //this.getOrPullModel(this.state.selectedModel);
-    });
-  };
-
-  sendPromptToServer = async () => {
-    this.setState({tableAIProgressLoading:true});
-
-
-    const serverEndpoint = "http://localhost:8000" + '/taxtech/tax_llm_url/';
+    const serverEndpoint = "http://localhost:8000/taxtech/tax_llm_url/";
     console.log(serverEndpoint);
 
-    const model = this.state.models.get(this.state.textSelectedModel);
-        
+    const model = models.get(textSelectedModel);
+
     const requestBody = {
-      model: model,
-      messages: [
-        {
-          role: 'user',
-          prompt: this.state.textprompt,
-        }
-      ],
+      model,
+      messages: [{ role: 'user', prompt: textPrompt }],
       stream: false,
-      isOnline: this.isOnline
+      isOnline,
     };
+
     try {
       const response = await axios.post(serverEndpoint, requestBody);
       const messageContent = response.data.response;
-      this.setState({tableAIProgressLoading:false});
-    
-      this.setState({ textresponse: messageContent });
-
-      return messageContent;
+      setTableAIProgressLoading(false);
+      setTextResponse(messageContent);
     } catch (error) {
-      console.error('Error processing Text Prompt:', (error as AxiosError).message);
-      this.setState({tableAIProgressLoading:false});
-      throw error;
+      console.error('Error processing Text Prompt:', error);
+      setTableAIProgressLoading(false);
     }
-    
   };
 
-  render(){
+  const handleTextPromptChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    setTextPrompt(event.target.value);
+  };
+
+  const handleTextModelChange = (event: SelectChangeEvent<string>) => {
+    setTextSelectedModel(event.target.value);
+  };
+
   return (
     <>
       <Box className="app-container">
@@ -129,54 +193,79 @@ class TaxTechDemo extends Component<TaxTechDemoProps, AppState> {
           <Divider />
           <Box className="input-container">
             <TextField
-              value={this.state.textprompt}
-              onChange={this.handleTextPromptChange}
+              value={textPrompt}
+              onChange={handleTextPromptChange}
               placeholder="Enter your prompt here..."
               fullWidth
               sx={{ backgroundColor: 'white', color: 'black' }}
             />
-            <Button
-              variant="contained"
-              onClick={this.sendPromptToServer}
-              disabled={this.state.isLoading}
-            >
-              {this.state.isLoading ? 'Processing...' : 'Submit'}
+            <Button variant="contained" onClick={sendPromptToServer} disabled={isLoading}>
+              {isLoading ? 'Processing...' : 'Submit'}
             </Button>
-            <Select
-              value={this.state.textSelectedModel}
-              onChange={this.handleTextModelChange}
-            >
-              {Array.from(this.state.models.entries()).map(([key, ]) => (
+            <Select value={textSelectedModel} onChange={handleTextModelChange}>
+              {Array.from(models.entries()).map(([key]) => (
                 <MenuItem key={key} value={key}>
                   {key}
                 </MenuItem>
               ))}
             </Select>
           </Box>
-          <Box id="botResult">
-            {this.state.tableAIProgressLoading && <LinearProgress />}
-          </Box>
-          {this.state.textresponse && (
+          <Box id="botResult">{tableAIProgressLoading && <LinearProgress />}</Box>
+          {textResponse && (
             <Box className="response-container">
               <h4>Response:</h4>
-              <TextField
-                value={JSON.stringify(this.state.textresponse, null, 2)}
-                disabled
-                multiline
-                fullWidth
-                rows={4}
-              />
+              <TextField value={JSON.stringify(textResponse, null, 2)} disabled multiline fullWidth rows={4} />
             </Box>
           )}
         </Box>
         <Divider sx={{ my: 2 }} />
       </Box>
-      <Grid item xs={12}>
-        <TaxDataDashboard />
-      </Grid>
+      <Box sx={{height: '100%'}}>
+      <Typography variant="h6" gutterBottom>
+        Appointment
+      </Typography>
+      <LocalizationProvider dateAdapter={AdapterDayjs}>
+      <DatePicker
+        label="Start Date"
+        value={startDate}
+        onChange={(newValue) => {
+          setStartDate(newValue);
+        }}
+      />
+      <DatePicker
+        label="End Date"
+        value={endDate}
+        onChange={(newValue) => {
+          setEndDate(newValue);
+        }}
+      />
+    </LocalizationProvider>
+    <DataGrid
+        rows={timerows}
+        columns={columns}
+        slots={{
+          toolbar: CustomToolbar,
+        }}
+        slotProps={{
+          toolbar: {
+            csvOptions: {allColumns: true, fileName: 'gridData'},
+          },
+        }}
+        initialState={{
+          pagination: {
+            paginationModel: {
+              pageSize: 10,
+              page: 0,
+            },
+          },
+        }}
+        pageSizeOptions={[10, 25, 50]}
+        checkboxSelection
+        disableRowSelectionOnClick
+      />
+    </Box>
     </>
-  )
-}
-}
+  );
+};
 
 export default TaxTechDemo;
