+++ frontend/src/reducers/tax/TaxDashboardDataReducer.tsx
@@ -1,6 +1,7 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit';
 
-const API_URL = import.meta.env.VITE_SANJEEVINI_BACKEND_APP_API_URL;
+const API_URL = "http://localhost:8000/";
+//import.meta.env.VITE_SANJEEVINI_BACKEND_APP_API_URL;
 
 export const fetchTaxDashboardData = createAsyncThunk<
 string[],
