+++ frontend/tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/App.tsx","./src/ErrorBoundary.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/assets/discordIcon.tsx","./src/components/About.tsx","./src/components/Demos.tsx","./src/components/Home.tsx","./src/components/Maps.tsx","./src/components/NavBar.tsx","./src/components/NoMatch.tsx","./src/components/Space.tsx","./src/components/chatbot/chatApp.tsx","./src/components/dashboard/Dashboard.tsx","./src/components/demos/DataTable.tsx","./src/components/demos/FoodGuardian.tsx","./src/components/demos/IndeterminateProgressBar.tsx","./src/components/demos/Recipes.tsx","./src/components/demos/SpeechASR.tsx","./src/components/demos/SpeechLLM.tsx","./src/components/demos/ollamaservice.ts","./src/components/demos/taxtech/TaxDataDashboard.tsx","./src/components/demos/taxtech/TaxDemo.tsx","./src/components/indic_llm/IndicDemo.tsx","./src/components/login/Login.tsx","./src/components/login/LoginActions.tsx","./src/components/login/LoginReducer.tsx","./src/components/login/LoginTypes.tsx","./src/components/news/News.tsx","./src/components/signup/Signup.tsx","./src/components/signup/SignupActions.tsx","./src/components/signup/SignupReducer.tsx","./src/components/signup/SignupTypes.tsx","./src/components/speech/SpeechDemo.tsx","./src/components/text_llm/TextDemo.tsx","./src/components/translate/TranslateDemo.tsx","./src/components/vision/VisionDemo.tsx","./src/reducers/Reducer.tsx","./src/reducers/Root.tsx","./src/reducers/store.tsx","./src/reducers/tax/TaxDashboardDataReducer.tsx","./src/utils/Utils.tsx"],"version":"5.6.2"}
+{"root":["./src/App.tsx","./src/ErrorBoundary.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/assets/discordIcon.tsx","./src/components/About.tsx","./src/components/Demos.tsx","./src/components/Home.tsx","./src/components/Maps.tsx","./src/components/NavBar.tsx","./src/components/NoMatch.tsx","./src/components/Space.tsx","./src/components/chatbot/chatApp.tsx","./src/components/dashboard/Dashboard.tsx","./src/components/demos/DataTable.tsx","./src/components/demos/FoodGuardian.tsx","./src/components/demos/IndeterminateProgressBar.tsx","./src/components/demos/Recipes.tsx","./src/components/demos/SpeechASR.tsx","./src/components/demos/SpeechLLM.tsx","./src/components/demos/ollamaservice.ts","./src/components/demos/taxtech/TaxDemo.tsx","./src/components/indic_llm/IndicDemo.tsx","./src/components/login/Login.tsx","./src/components/login/LoginActions.tsx","./src/components/login/LoginReducer.tsx","./src/components/login/LoginTypes.tsx","./src/components/news/News.tsx","./src/components/signup/Signup.tsx","./src/components/signup/SignupActions.tsx","./src/components/signup/SignupReducer.tsx","./src/components/signup/SignupTypes.tsx","./src/components/speech/SpeechDemo.tsx","./src/components/text_llm/TextDemo.tsx","./src/components/translate/TranslateDemo.tsx","./src/components/vision/VisionDemo.tsx","./src/reducers/Reducer.tsx","./src/reducers/Root.tsx","./src/reducers/store.tsx","./src/reducers/tax/TaxDashboardDataReducer.tsx","./src/utils/Utils.tsx"],"errors":true,"version":"5.6.2"}
